pyproject.toml
[project]
name = "kurs-lokalny"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "ipykernel>=7.2.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.12",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.49",
]
